+++ src/test/java/jewellery/inventory/unit/mapper/SaleMapperTest.java
@@ -5,6 +5,7 @@ import static jewellery.inventory.helper.UserTestHelper.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import jewellery.inventory.dto.request.ProductPriceDiscountRequestDto;
@@ -53,7 +54,7 @@ class SaleMapperTest {
     productsForSale = SaleTestHelper.getProductsList(product);
     sale = SaleTestHelper.createSaleWithTodayDate(seller, buyer, productsForSale);
     productPriceDiscountRequestDto =
-        SaleTestHelper.createProductPriceDiscountRequest(product.getId(), 1000, 10);
+        SaleTestHelper.createProductPriceDiscountRequest(product.getId(), BigDecimal.valueOf(1000), BigDecimal.valueOf(10));
     List<ProductPriceDiscountRequestDto> productPriceDiscountRequestDtoList = new ArrayList<>();
     productPriceDiscountRequestDtoList.add(productPriceDiscountRequestDto);
     saleRequestDto =
@@ -93,7 +94,7 @@ class SaleMapperTest {
     when(userMapper.toUserResponse(buyer)).thenReturn(buyerResponseDto);
 
     when(productMapper.mapToProductResponseDto(product)).thenReturn(new ProductResponseDto());
-    sale.getProducts().get(0).setSalePrice(0);
+    sale.getProducts().get(0).setSalePrice(BigDecimal.valueOf(0));
     assertThrows(
             IllegalArgumentException.class, () -> saleMapper.mapEntityToResponseDto(sale));
    }
